refactor: extract trial hp table to be reusable
