+++ webui/react/src/pages/ExperimentDetails/ExperimentVisualization/LearningCurve.module.scss
@@ -8,16 +8,6 @@
   .chart {
     height: 40rem;
   }
-  .idLayout {
-    align-items: center;
-    display: flex;
-  }
-  .colorLegend {
-    border-radius: var(--theme-sizes-border-radius);
-    height: 2rem;
-    margin-right: var(--theme-sizes-layout-medium);
-    width: 2rem;
-  }
   .table {
     padding: var(--theme-sizes-layout-big);
   }
